// example/convex/_generated/api.d.ts
declare const fullApi: ApiFromModules<{
  autosendDemo: typeof autosendDemo;
  crons: typeof crons;
  email: typeof email;
  envSetup: typeof envSetup;
  http: typeof http;
  mailtm: typeof mailtm;
}>;

/**
 * A utility for referencing Convex functions in your app's public API.
 *
 * Usage:
 * ```js
 * const myFunctionReference = api.myModule.myFunction;
 * ```
 */
export declare const api: FilterApi<
  typeof fullApi,
  FunctionReference<any, "public">
>;